requirements.txt
asgiref==3.7.2
Django==3.2.23
djangorestframework==3.14.0
dnspython==2.3.0
mongoengine==0.27.0
pymongo==4.6.1
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.7.1
